requirements.txt
matplotlib==2.2.2
torchvision==0.4.1
torch==1.3.0
Pillow==6.2.1
